Added all the abilities and types with .joined(). Updated the DetailVC to populate all the fields, but it needs some beuatifying and debugging
