--- Pokedex/Pokedex/ViewControllers/DetailViewController.swift
@@ -1,36 +1,0 @@
-//
-//  DetailViewController.swift
-//  Pokedex
-//
-//  Created by Chris Price on 3/21/20.
-//  Copyright © 2020 com.chrispriiice. All rights reserved.
-//
-
-import UIKit
-
-class DetailViewController: UIViewController {
-
-    @IBOutlet weak var pokemonNameLabel: UILabel!
-    @IBOutlet weak var pokemonIDLabel: UILabel!
-    @IBOutlet weak var pokemonTypesLabel: UILabel!
-    @IBOutlet weak var pokemonAbilitiesLabel: UILabel!
-    @IBOutlet weak var pokemonImageView: UIImageView!
-    
-    var pokemon: Pokemon? {
-        didSet {
-                self.updateViews()
-        }
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        updateViews()
-        title = pokemon?.name.capitalized
-    }
-    
-    func updateViews() {
-        guard let pokemon = pokemon else { return }
-        pokemonNameLabel.text = pokemon.name.capitalized
-        pokemonIDLabel.text = "\(pokemon.id)"
-    }
-}
