+++ Pokedex/Resources/Assets.xcassets/Contents.json
