+++ Pokedex/Resources/Info.plist
