+++ Pokedex/ViewControllers/SearchViewController.swift
@@ -17,6 +17,7 @@ class SearchViewController: UIViewController {
     @IBOutlet weak var pokemonAbilitiesLabel: UILabel!
     @IBOutlet weak var pokemonImageView: UIImageView!
     
+    
     var pokemonController: PokemonController?
     var pokemon: Pokemon? {
         didSet {
@@ -32,12 +33,10 @@ class SearchViewController: UIViewController {
     
     func updateViews() {
         guard let pokemon = pokemon else { return }
-        print(pokemon.sprites.frontDefault)
         pokemonNameLabel.text = pokemon.name.capitalized
         pokemonIDLabel.text = "\(pokemon.id)"
-        pokemonAbilitiesLabel.text = pokemon.abilities.first?.ability.name
-        pokemonTypesLabel.text = pokemon.types.first?.type.name
-        // Set the image to the pokemon sprite
+        pokemonAbilitiesLabel.text = pokemon.abilities.map({$0.ability.name}).joined(separator: ", ")
+        pokemonTypesLabel.text = pokemon.types.map({$0.type.name}).joined(separator: ", ")
         pokemonController?.fetchImage(urlString: pokemon.sprites.frontDefault, completion: { (result) in
             if let pokemonSearchResult = try? result.get() {
                 DispatchQueue.main.async {
@@ -61,7 +60,6 @@ extension SearchViewController: UISearchBarDelegate {
             if let pokemonSearchResult = try? result.get() {
                 DispatchQueue.main.async {
                     self.pokemon = pokemonSearchResult
-                    print(self.pokemon)
                 }
             }
         })
